Modify download_from_s3 to correctly fall back to basename if frameid is not present. Add tests for get_basename method.
